refactor(desktop): inline preset auto-add as a plain function
