--- apps/desktop/src/renderer/routes/_authenticated/settings/agents/components/V2AgentsSettings/hooks/useAutoAddAgentPreset/index.ts
@@ -1,1 +1,0 @@
-export { useAutoAddAgentPreset } from "./useAutoAddAgentPreset";
